+++ src/containerapp/azext_containerapp/_models.py
@@ -546,3 +546,7 @@ JavaComponent = {
         "componentType": None
     }
 }
+
+RuntimeJava = {
+    "enableMetrics": False
+}