+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,8 @@ Release History
 ===============
 upcoming
 ++++++
+* Upgrade api-version to 2023-11-02-preview
+* 'az containerapp create/update': Support enable or disable Java metrics with --runtime and --enable-java-metrics
 
 0.3.48
 ++++++